+++ default.nix
@@ -63,7 +63,7 @@ let
       { name = "postgresql-12"; postgresql = pkgs.postgresql_12.withPackages (p: [ p.postgis p.pg_safeupdate ]); }
       { name = "postgresql-11"; postgresql = pkgs.postgresql_11.withPackages (p: [ p.postgis p.pg_safeupdate ]); }
       { name = "postgresql-10"; postgresql = pkgs.postgresql_10.withPackages (p: [ p.postgis p.pg_safeupdate ]); }
-      { name = "postgresql-9.6"; postgresql = pkgs.postgresql_9_6.withPackages (p: [ p.postgis p.pg_safeupdate ]); }
+      { name = "postgresql-9_6"; postgresql = pkgs.postgresql_9_6.withPackages (p: [ p.postgis p.pg_safeupdate ]); }
     ];
 
   # Dynamic derivation for PostgREST
@@ -77,7 +77,7 @@ let
     "-f dev --test-show-detail=direct";
 
   profiledHaskellPackages =
-    pkgs.haskell.packages."${compiler}".extend (self: super:
+    pkgs.haskell.packages."${compiler}".extend (_: super:
       {
         mkDerivation =
           args:
@@ -118,7 +118,7 @@ rec {
     pkgs.callPackage nix/tools/cabalTools.nix { inherit devCabalOptions postgrest; };
 
   withTools =
-    pkgs.callPackage nix/tools/withTools.nix { inherit cabalTools devCabalOptions postgresqlVersions postgrest; };
+    pkgs.callPackage nix/tools/withTools.nix { inherit postgresqlVersions postgrest; };
 
   # Development tools.
   devTools =